separate test for full text
